made changes to refactor
