+++ dbc_human.rb
@@ -1,0 +1,14 @@
+class Human
+  attr_reader :age, :phase
+  attr_accessor :name
+
+  def offer_high_five
+    "High five!"
+  end
+
+
+
+end
+
+
+
