+++ student.rb
@@ -1,15 +1,11 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'dbc_human'
+
+class Student < Human
 
   def initialize(options = {})
     @phase = 1
     @age = options.fetch(:age, 0)
     @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
@@ -31,3 +27,6 @@ class Student
     response
   end
 end
+
+stud = Student.new
+p stud.class
